second app

+++ javaschool/src/main/java/com/tsystems/service/impl/TruckServiceImpl.java
@@ -15,10 +15,15 @@ import com.tsystems.entity.UserOrder;
 import com.tsystems.enumaration.DriverStatus;
 import com.tsystems.enumaration.TruckStatus;
 import com.tsystems.enumaration.UserOrderStatus;
+import com.tsystems.exception.DataChangingException;
 import com.tsystems.service.api.CountingService;
+import com.tsystems.service.api.JMSSenderService;
 import com.tsystems.service.api.TruckService;
+import com.tsystems.utils.TruckPair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.joda.time.DateTime;
+import org.joda.time.Days;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -51,10 +56,13 @@ public class TruckServiceImpl implements TruckService {
     @Autowired
     private CountingService countingService;
 
+    @Autowired
+    private JMSSenderService jmsSenderService;
+
     /**
      * Saves a new truck
      *
-     * @param truckDto truck
+     * @param truckDto - truck
      * @return true if truck was added
      */
     @Override
@@ -70,6 +78,7 @@ public class TruckServiceImpl implements TruckService {
             truck.setLongitude(office.getLongitude());
             truckDao.save(truck);
             LOGGER.info("Added a new truck with registration number " + truck.getRegistrationNumber());
+            jmsSenderService.sendMessage();
             return true;
         }
         return false;
@@ -78,7 +87,7 @@ public class TruckServiceImpl implements TruckService {
     /**
      * Updates the truck
      *
-     * @param truckDto truck
+     * @param truckDto - truck
      * @return true if truck was updated
      */
     @Override
@@ -98,22 +107,26 @@ public class TruckServiceImpl implements TruckService {
     /**
      * Delete truck by id
      *
-     * @param id truck id
+     * @param id - truck id
      * @return true if truck was deleted
      */
     @Override
     @Transactional
     public boolean deleteById(long id) {
         Truck truck = truckDao.findById(id);
+        if (truck.getUserOrder() != null || !truck.getDriverList().isEmpty()) {
+            throw new DataChangingException("Cant delete this truck");
+        }
         truckDao.delete(truck);
         LOGGER.info("Deleted a truck with registration number " + truck.getRegistrationNumber());
+        jmsSenderService.sendMessage();
         return true;
     }
 
     /**
      * Finds a truck by id
      *
-     * @param id truck id
+     * @param id - truck id
      * @return TruckDto
      */
     @Override
@@ -125,28 +138,36 @@ public class TruckServiceImpl implements TruckService {
     /**
      * Finds all available trucks for this order and sort and sorts them in order of increasing distance
      *
-     * @param orderId order id
+     * @param orderId - order id
      * @return List<TruckDto>
      */
     @Override
     @Transactional
     public List<TruckDto> findAllAvailable(long orderId) {
-        UserOrder userOrder=userOrderDao.findById(orderId);
+        UserOrder userOrder = userOrderDao.findById(orderId);
         List<Truck> trucks = truckDao.findAllAvailable();
-        Predicate<Truck> haveOrder = truck -> truck.getUserOrder() == null;
+        Predicate<Truck> haveOrder = truck -> truck.getUserOrder() != null;
         trucks.removeIf(haveOrder);
-        List<TruckDto> truckDtoList=trucks.stream()
+        List<TruckDto> truckDtoList = trucks.stream()
                 .map(truck -> modelMapper.map(truck, TruckDto.class))
                 .collect(Collectors.toList());
-        return (List<TruckDto>) countingService.getApproximatelyTotalDistanceForTruckAndOrder(truckDtoList,
-                modelMapper.map(userOrder, UserOrderDto.class)).values();
+        List<TruckPair> truckPairs = countingService.getApproximatelyTotalDistanceForTruckAndOrder(truckDtoList,
+                modelMapper.map(userOrder, UserOrderDto.class));
+
+        //checking that the time limit of 176 hours per month will not be exceeded for any driver
+        Predicate<TruckPair> isLimitForDriversExceeded = this::isLimitForDriversExceeded;
+        truckPairs.removeIf(isLimitForDriversExceeded);
+        Collections.sort(truckPairs);
+        return truckPairs.stream()
+                .map(TruckPair::getTruckDto)
+                .collect(Collectors.toList());
     }
 
     /**
      * Adds order to truck
      *
-     * @param truckId truck id
-     * @param orderId order id
+     * @param truckId - truck id
+     * @param orderId - order id
      * @return true if order was added
      */
     @Override
@@ -154,18 +175,24 @@ public class TruckServiceImpl implements TruckService {
     public boolean addOrder(long truckId, long orderId) {
         Truck truck = truckDao.findById(truckId);
         UserOrder userOrder = userOrderDao.findById(orderId);
-        if (userOrder.getStatus().equals(UserOrderStatus.NOT_TAKEN) && !userOrder.getCargoList().isEmpty()) {
-            userOrder.setTruck(truck);
-            userOrder.setStatus(UserOrderStatus.TAKEN);
-            return true;
+        if (!userOrder.getStatus().equals(UserOrderStatus.NOT_TAKEN) ||
+                userOrder.getCargoList().isEmpty() ||
+                truck.getStatus().equals(TruckStatus.FAULTY) ||
+                truck.getUserOrder() != null) {
+            throw new DataChangingException("Cant add this truck to this order");
         }
-        return false;
+        userOrder.setTruck(truck);
+        userOrder.setStatus(UserOrderStatus.TAKEN);
+        LOGGER.info("The truck with the number " + truck.getRegistrationNumber() +
+                " was assigned to the order with the number " + userOrder.getUniqueNumber());
+        jmsSenderService.sendMessage();
+        return true;
     }
 
     /**
      * Finds all available drivers for this truck
      *
-     * @param truckId truck id
+     * @param truckId - truck id
      * @return List<DriverDto>
      */
     @Override
@@ -187,8 +214,8 @@ public class TruckServiceImpl implements TruckService {
     /**
      * Adds driver to the truck
      *
-     * @param truckId  truck id
-     * @param driverId driver id
+     * @param truckId  - truck id
+     * @param driverId - driver id
      * @return true if driver was added
      */
     @Override
@@ -198,18 +225,20 @@ public class TruckServiceImpl implements TruckService {
         Truck truck = truckDao.findById(truckId);
 
         //the truck does not fulfill the order and the driver does not have a car
-        if (truck.getUserOrder() == null && driver.getTruck() == null) {
-            driver.setTruck(truck);
-            return true;
+        if (truck.getUserOrder() != null || driver.getTruck() != null) {
+            throw new DataChangingException("Cant add this driver to this truck");
         }
-        return false;
+        driver.setTruck(truck);
+        LOGGER.info("The driver with the number " + driver.getPersonalNumber() +
+                " was assigned to the truck with the registration number " + truck.getRegistrationNumber());
+        return true;
     }
 
     /**
      * Removes the driver from the truck
      *
-     * @param truckId  truck id
-     * @param driverId driver id
+     * @param truckId  - truck id
+     * @param driverId - driver id
      * @return true if driver was removed
      */
     @Override
@@ -217,165 +246,59 @@ public class TruckServiceImpl implements TruckService {
     public boolean removeDriver(long truckId, long driverId) {
         Truck truck = truckDao.findById(truckId);
         Driver driver = driverDao.findById(driverId);
-        if (truck.getUserOrder() == null) {
-
-            //change driver status to rest
-            if (!driver.getStatus().equals(DriverStatus.REST)) {
-                double hoursThisMonth = driver.getHoursThisMonth();
-                double hoursWorked = countingService.getDriverHours(driver);
-                driver.setHoursThisMonth(hoursThisMonth + hoursWorked);
-                driver.setStatus(DriverStatus.REST);
-            }
-            driver.setTruck(null);
-            return true;
+        if (truck.getUserOrder() != null) {
+            throw new DataChangingException("Cant remove this driver because this driver is carrying an order");
         }
-        return false;
+
+        //change driver status to rest
+        if (!driver.getStatus().equals(DriverStatus.REST)) {
+            double hoursThisMonth = driver.getHoursThisMonth();
+            double hoursWorked = countingService.getDriverHours(driver);
+            driver.setHoursThisMonth(hoursThisMonth + hoursWorked);
+            driver.setStatus(DriverStatus.REST);
+            LOGGER.info("The status of the driver with the number " +
+                    driver.getPersonalNumber() + " was changed to rest");
+        }
+        driver.setTruck(null);
+        LOGGER.info("The driver with the number " + driver.getPersonalNumber() +
+                " was removed from the truck with the registration number " + truck.getRegistrationNumber());
+        return true;
     }
 
-//    /**
-//     * Finds all available trucks for this order and sort and sorts them in order of increasing distance
-//     *
-//     * @param userOrderDto
-//     * @return List<TruckDto>
-//     */
-//    @Override
-//    @Transactional
-//    public List<TruckDto> findAllAvailable(UserOrderDto userOrderDto) {
-//        if (!(userOrderDto.getTruck() == null)) {
-//            return null;
-//        }
-//        List<Truck> trucks = truckDao.findAll();
-//        Iterator<Truck> iterator = trucks.iterator();
-//        Truck truck;
-//        while (iterator.hasNext()) {
-//            truck = iterator.next();
-//
-//            //the condition of the truck should be normal
-//            if (truck.getStatus().equals(TruckStatus.FAULTY)) {
-//                iterator.remove();
-//                continue;
-//            }
-//
-//            //the truck must not have any other orders
-//            if (!truck.getUserOrderList().isEmpty()) {
-//                iterator.remove();
-//                continue;
-//            }
-//
-//            //the truck must have at least 1 driver
-//            if (truck.getDriverList().isEmpty()) {
-//                iterator.remove();
-//                continue;
-//            }
-//
-//            //the load must not exceed the load capacity of the truck
-//            if (truck.getWeightCapacity() < getTotalWeight(userOrderDto.getCargoList())) {
-//                iterator.remove();
-//                continue;
-//            }
-//
-//            //checking that the time limit of 176 hours per month will not be exceeded for any driver
-//            if (isLimitForDriversExceeded(truck, userOrderDto)) {
-//                iterator.remove();
-//            }
-//        }
-//        if (trucks.isEmpty()) {
-//            return null;
-//        }
-//
-//        //sorting by distance
-//        trucks = sortByDistance(trucks, userOrderDto.getWaypointList().get(0).getLocation());
-//        return trucks.stream()
-//                .map(this::convertToDto)
-//                .collect(Collectors.toList());
-//    }
+    /**
+     * Checks that the drivers is approach for this order
+     *
+     * @param truckPair - truck and distance for this truck and order
+     * @return true if one of the drivers has exceeded the shift limit
+     */
+    private boolean isLimitForDriversExceeded(TruckPair truckPair) {
+        //the speed of the truck in kilometers per hour
+        final double truckSpeed = 90.0;
+        final double maxHoursPerMonth = 176.0;
+        final double hoursForOrder = (float) (truckPair.getApproximatelyTotalDistanceForTruckAndOrder() / truckSpeed);
+        double driverHoursThisMonth;
+        for (DriverDto driverDto : truckPair.getTruckDto().getDriverList()) {
+            driverHoursThisMonth = driverDto.getHoursThisMonth();
+            if ((driverHoursThisMonth + hoursForOrder) < maxHoursPerMonth) {
+                return false;
+            } else {
+                DateTime currentDate = new DateTime();
+                int month = currentDate.getMonthOfYear() + 1;
+                int year = currentDate.getYear();
+                if (month == 13) {
+                    month = 1;
+                    year++;
+                }
+                DateTime nextMonth = new DateTime(year, month, 1, 0, 0);
+                Days days = Days.daysBetween(currentDate, nextMonth);
+                double daysToEndThisMonth = days.getDays();
 
-//    /**
-//     * Calculate total weight for order
-//     *
-//     * @param cargoDtoList
-//     * @return weight in kilograms
-//     */
-//    private double getTotalWeight(List<CargoDto> cargoDtoList) {
-//        double totalWeight = 0.0;
-//        for (CargoDto cargoDto : cargoDtoList) {
-//            totalWeight += cargoDto.getWeight();
-//        }
-//        return totalWeight;
-//    }
-//
-//    /**
-//     * Sorts the list of trucks by truck location and order place of departure
-//     *
-//     * @param trucks
-//     * @param locationDto
-//     * @return List<Truck>
-//     */
-//    private List<Truck> sortByDistance(List<Truck> trucks, LocationDto locationDto) {
-//        TreeMap<Integer, Truck> truckDistanceTreeMap = new TreeMap<>();
-//        int distance;
-//        for (Truck truck : trucks) {
-//            distance = countingService.getDistanceLength(truck.getLocation().getLatitude(),
-//                    truck.getLocation().getLongitude(),
-//                    locationDto.getLatitude(),
-//                    locationDto.getLongitude());
-//            truckDistanceTreeMap.put(distance, truck);
-//        }
-//        return new ArrayList<>(truckDistanceTreeMap.values());
-//    }
-//
-//    /**
-//     * Checks that the drivers is approach for this order
-//     *
-//     * @param truck
-//     * @param userOrderDto
-//     * @return true if one of the drivers has exceeded the shift limit
-//     */
-//    private boolean isLimitForDriversExceeded(Truck truck, UserOrderDto userOrderDto) {
-//        //the speed of the truck in kilometers per hour
-//        double truckSpeed = 40.0;
-//        double maxHoursPerMonth = 176.0;
-//        double distanceBetweenTruckAndOrder = getDistanceBetweenTruckAndOrder(truck, userOrderDto);
-//        double hoursForOrder = (float) ((userOrderDto.getDistance() + distanceBetweenTruckAndOrder) / truckSpeed);
-//        List<Driver> drivers = truck.getDriverList();
-//        double driverHoursThisMonth;
-//        for (Driver driver : drivers) {
-//            driverHoursThisMonth = driver.getHoursThisMonth();
-//            if ((driverHoursThisMonth + hoursForOrder) < maxHoursPerMonth) {
-//                return false;
-//
-//                //checking if the month changes
-//            } else {
-//                DateTime currentDate = new DateTime();
-//                int month = currentDate.getMonthOfYear() + 1;
-//                int year = currentDate.getYear();
-//                if (month == 13) {
-//                    month = 1;
-//                    year++;
-//                }
-//                DateTime nextMonth = new DateTime(year, month, 1, 0, 0);
-//                Days days = Days.daysBetween(currentDate, nextMonth);
-//                double daysToEndThisMonth = days.getDays();
-//
-//                //even if the driver goes out to order every day
-//                double maxHoursInThisMonth = daysToEndThisMonth * truck.getDriverShiftSize();
-//                return (driverHoursThisMonth + maxHoursInThisMonth) > maxHoursPerMonth;
-//            }
-//        }
-//        return true;
-//    }
-//
-//    /**
-//     * Calculates the distance between the truck and the order
-//     *
-//     * @param truck
-//     * @param userOrderDto
-//     * @return distance in km
-//     */
-//    private double getDistanceBetweenTruckAndOrder(Truck truck, UserOrderDto userOrderDto) {
-//        return countingService.getDistanceLength(truck.getLocation().getLatitude(),
-//                truck.getLocation().getLongitude(),
-//                userOrderDto.getWaypointList().get(0).getLocation().getLatitude(),
-//                userOrderDto.getWaypointList().get(0).getLocation().getLongitude());
-//    }
+                //even if the driver goes out to order every day
+                double maxHoursInThisMonth = daysToEndThisMonth * truckPair.getTruckDto().getDriverShiftSize();
+                return (driverHoursThisMonth + maxHoursInThisMonth) > maxHoursPerMonth;
+            }
+        }
+        return true;
+    }
 }
+
